test(routing): split the routing suite to match the layer split
`tests/test_model_routing.py` reached 1,491 lines and crossed the branch's own
1,000-line deep-quality gate — the production split landed without the suite
following it.

Split by AST span into `routing_fixtures.py` (shared ROOT, manifest, catalog,
stub runners, `RoutingTestCase`) plus one module per layer: policy, manifest,
closure, resolver, transport. `test_model_routing.py` keeps only the two facade
guards that are about the facade itself — layer ordering and re-export
completeness.

Verified a pure move: 47 test methods before, 47 after, none added, removed, or
changed in body. Modules import `from routing_fixtures import ...` rather than
`tests.routing_fixtures`, which resolves under both `pytest tests/` and
`unittest discover -s tests` (there is no `tests/__init__.py`).

Also carried here, since it predates the split: a negative lens-menu fixture
combined a duplicate path with a nonexistent one, so it never independently
tested path existence. Now two fixtures.
